requirements.txt
# Python 3.11.6
pygame
opencv-python
Pillow
tk

flask-cors